Added fileType attribute in ImagesGroup.cs to accomodate multiple images processors and now degree of similarity in SearchParameters.cs is an int.

+++ Core/Entities/SearchParameters.cs
@@ -11,7 +11,7 @@ public class SearchParameters
     public FileSearchType? FileSearchType { get; set; }
     
     // The degree of similarity between images (only for images)
-    public double? DegreeOfSimilarity { get; set; }
+    public int? DegreeOfSimilarity { get; set; }
 
     // Should we include subfolders or not
     public bool IncludeSubfolders { get; set; }
